LPS-188020 Handle ETag correctly in LanguageFilter
We need to modify the ETag returned by the delegate filter chain because
we are modifying the content.

Also, instead of caching the responses we cache ETags in a map indexed by
(file path, language id). This way, if the ETag sent from the browser
matches the latest one sent by the server (for a given path and language)
we return SC_NOT_MODIFIED. This way, we maintain the browser cache valid
until the server restarts.

Note also that, in case the server restarts and nothing has changed, only
the first client accessing a JS will be penalized by a cache invalidation
because, after that, the same ETag we used in the previous server restart
will be cached and all subsequent clients will get a SC_NOT_MODIFIED again.